requirements.txt
pytorch-lightning >= 1.0.0rc2
Pillow
defusedxml
rapidfuzz>=1.6.0
wandb>=0.10.31
fastprogress
pytest>=5.3.2
requests>=2.20.0
hdf5storage>=0.1.18
coverage>=4.5.4
requirements-parser==0.2.0
numpy>=1.16.0
scipy>=1.4.0
h5py>=3.1.0
opencv-python>=3.4.5.20
PyMuPDF>=1.16.0,!=1.18.11,!=1.18.12
pyclipper>=1.2.0
shapely>=1.6.0
matplotlib>=3.1.0,<3.4.3
mplcursors>=0.3
weasyprint>=52.2,<53.0
unidecode>=1.0.0
tqdm>=4.30.0
streamlit>=1.0.0
importlib_metadata
fastprogress>=0.1.21
